django/requirements.txt
Django==3.0.7
djangorestframework==3.11.0
requests==2.23.0
django-cors-headers==3.4.0